[UI] display env banner in integration and pre-prod

+++ frontend/src/features/MainWindow/components/BannerStack/index.tsx
@@ -13,6 +13,11 @@ export function BannerStack() {
   const bannerStack = useAppSelector(state => state.mainWindow.bannerStack)
 
   const remove = (bannerStackRank: number) => {
+    const bannerToRemove = bannerStack.entities[bannerStackRank]
+    if (bannerToRemove?.props.isCollapsible) {
+      return
+    }
+
     dispatch(mainWindowActions.removeBanner(bannerStackRank))
   }
